Lints

+++ store/src/server/api.rs
@@ -320,7 +320,7 @@ impl api_server::Api for StoreApi {
                         .map_err(|err: ConversionError| {
                             Status::invalid_argument(err.to_string())
                         })?,
-                    note_type: NoteType::try_from(note.note_type)
+                    note_type: NoteType::try_from(note.note_type as u64)
                         .map_err(|err: NoteError| Status::invalid_argument(err.to_string()))?,
                     sender: note.sender.ok_or(invalid_argument("Note missing sender"))?.into(),
                     tag: note.tag,
